\name{gdp4795}
\docType{data}
\alias{gdp4795}

\title{USA Real Gross Domestic Product}

\description{
Quarterly U.S. real gross domestic product for the period 1947:I-1995:III.
}

\usage{
gdp4795
}

\format{
A time series object.
}

%\source{internet
%Clark 1987, data are available from the author on request.}

\references{
Clark, P.K. (1987).
\sQuote{The Cyclical Component of U.S. Economic Activity},
\emph{Quarterly Journal of Economics}, \bold{102}(4), pp. 797-814.
}

\keyword{datasets}
